feat: Create a new landing page and refactor the application
This commit introduces a new landing page for the GreenTrack project, as per the user's request. The landing page provides information about the application, its features, and a contact form.

The main application has been refactored into a modular structure with separate HTML, CSS, and JavaScript files to improve maintainability. The JavaScript code has been broken down into smaller, more manageable modules.

The theme of the main application has been updated to a light theme to match the new landing page, providing a consistent user experience.

The project structure has been cleaned up by removing the old, monolithic `index.html` file and renaming the new main application file to `index.html`.
