--- GreenTrack/index_new.html
@@ -1,142 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>GreenTrack - Digital Waste Management Platform</title>
-
-    <!-- PWA Manifest -->
-    <link rel="manifest" href="manifest.json">
-    <meta name="theme-color" content="#2d5a27">
-
-    <!-- Leaflet CSS -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.9.4/leaflet.min.css">
-
-    <!-- Chart.js -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/3.9.1/chart.min.js"></script>
-
-    <!-- Leaflet JS -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.9.4/leaflet.min.js"></script>
-
-    <!-- Styles -->
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <noscript>
-        <div style="padding: 2rem; text-align: center;">
-            <h1>JavaScript Required</h1>
-            <p>This application requires JavaScript to function properly. Please enable JavaScript in your browser.</p>
-        </div>
-    </noscript>
-
-    <a href="#main-content" class="skip-to-content">Skip to main content</a>
-
-    <header class="header">
-        <div class="header-content">
-            <div class="logo">
-                <span>🌱</span>
-                <span>GreenTrack</span>
-            </div>
-            <nav class="nav" id="nav">
-                <!-- Navigation will be populated by JavaScript -->
-            </nav>
-        </div>
-    </header>
-
-    <main class="main-content" id="main-content">
-        <div id="app">
-            <!-- App content will be populated by JavaScript -->
-        </div>
-    </main>
-
-    <!-- Toast container -->
-    <div id="toast" class="toast"></div>
-
-    <!-- Modal container -->
-    <div id="modal" class="modal">
-        <div class="modal-content" id="modal-content">
-            <!-- Modal content will be populated by JavaScript -->
-        </div>
-    </div>
-
-    <script src="app.js"></script>
-
-    <!-- Service Worker Script -->
-    <script id="sw-script" type="application/javascript">
-        const CACHE_NAME = 'greentrack-v1';
-        const urlsToCache = [
-            '/',
-            '/index.html',
-            '/css/styles.css',
-            '/js/app.js',
-            'https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.9.4/leaflet.min.css',
-            'https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.9.4/leaflet.min.js',
-            'https://cdnjs.cloudflare.com/ajax/libs/Chart.js/3.9.1/chart.min.js'
-        ];
-
-        self.addEventListener('install', event => {
-            event.waitUntil(
-                caches.open(CACHE_NAME)
-                    .then(cache => cache.addAll(urlsToCache))
-            );
-        });
-
-        self.addEventListener('fetch', event => {
-            event.respondWith(
-                caches.match(event.request)
-                    .then(response => {
-                        return response || fetch(event.request);
-                    })
-            );
-        });
-    </script>
-
-    <!-- PWA Manifest as inline JSON -->
-    <script>
-        // Create manifest.json content
-        const manifest = {
-            "name": "GreenTrack - Digital Waste Management",
-            "short_name": "GreenTrack",
-            "description": "Digital waste management and training platform",
-            "start_url": "/",
-            "display": "standalone",
-            "background_color": "#2d5a27",
-            "theme_color": "#2d5a27",
-            "icons": [
-                {
-                    "src": "data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTkyIiBoZWlnaHQ9IjE5MiIgdmlld0JveD0iMCAwIDE5MiAxOTIiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+CjxyZWN0IHdpZHRoPSIxOTIiIGhlaWdodD0iMTkyIiByeD0iMjQiIGZpbGw9IiMyZDVhMjciLz4KPHN2ZyB4PSI0OCIgeT0iNDgiIHdpZHRoPSI5NiIgaGVpZ2h0PSI5NiIgdmlld0JveD0iMCAwIDI0IDI0IiBmaWxsPSIjZmZmZmZmIj4KPHA+8J+MsTwvcD4KPC9zdmc+Cjwvc3ZnPgo=",
-                    "sizes": "192x192",
-                    "type": "image/svg+xml"
-                },
-                {
-                    "src": "data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNTEyIiBoZWlnaHQ9IjUxMiIgdmlld0JveD0iMCAwIDUxMiA1MTIiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+CjxyZWN0IHdpZHRoPSI1MTIiIGhlaWdodD0iNTEyIiByeD0iNjQiIGZpbGw9IiMyZDVhMjciLz4KPHN2ZyB4PSIxMjgiIHk9IjEyOCIgd2lkdGg9IjI1NiIgaGVpZ2h0PSIyNTYiIHZpZXdCb3g9IjAgMCAyNCAyNCIgZmlsbD0iI2ZmZmZmZiI+CjxwPvCfjLE8L3A+Cjwvc3ZnPgo8L3N2Zz4K",
-                    "sizes": "512x512",
-                    "type": "image/svg+xml"
-                }
-            ]
-        };
-
-        // Create and inject manifest
-        const manifestBlob = new Blob([JSON.stringify(manifest)], {type: 'application/json'});
-        const manifestURL = URL.createObjectURL(manifestBlob);
-        document.querySelector('link[rel="manifest"]').href = manifestURL;
-
-        // Create and inject service worker
-        const swScript = document.getElementById('sw-script').textContent;
-        const swBlob = new Blob([swScript], {type: 'application/javascript'});
-        const swURL = URL.createObjectURL(swBlob);
-
-        if ('serviceWorker' in navigator) {
-            window.addEventListener('load', () => {
-                navigator.serviceWorker.register(swURL)
-                    .then(registration => {
-                        console.log('SW registered: ', registration);
-                    })
-                    .catch(registrationError => {
-                        console.log('SW registration failed: ', registrationError);
-                    });
-            });
-        }
-    </script>
-</body>
-</html>
